move snap fuzzers
